(Refactor) Put date formatting into a helper to DRY up the controllers
