+++ app/helpers/date_helper.rb
@@ -1,0 +1,11 @@
+module DateHelper
+  def format_date(params)
+    date_parts = params.values_at(:day, :month, :year)
+    return unless date_parts.all?(&:present?)
+
+    day, month, year = date_parts.map(&:to_i)
+    Date.new(year, month, day)
+  rescue ArgumentError
+    nil
+  end
+end
